Multiple colors in shortcode attributes for backgrounds and borders
Shortcode attributes `backgroundColor` and `borderColor` are renamed `backgroundColors` and `borderColors` respectively. Note that singular attributes are still supported as synonyms of plural attributes.

The attribute `backgroundColors` can contain multiple colors separated by comma. If some colors are missing (for example, if we have 4 datasets and only 1 color is specified in the attribute), default colors from the theme (attribute `theme`) are used as fallback.

The attribute `borderColors` can contain multiple colors separated by comma. If some colors are missing, colors from the attribute `backgroundColors` are used as fallback if they're specified, otherwise default colors from the theme are used.

For example, the following shortcode specifies only 1 background color for the second dataset, and it specifies border colors for the first dataset and the third one:

    [chartjs type="bar" barThickness="25" theme="Visualizer" backgroundColors=",#FFFF00" borderColors="#0000FF,,#008000"]Year,Dataset1,Dataset2,Dataset3,Dataset4
    2018,50,25,32,45
    2019,32,45,22,19
    2020,25,49,35,23
    2021,42,31,43,35[/chartjs]

The shortcode attributes are as follows:

| Color attributes     | Dataset1 | Dataset2 | Dataset3 | Dataset4 |
|----------------------|----------|----------|----------|----------|
| **backgroundColors** | empty    | #FFFF00  | empty    | empty    |
| **borderColors**     | #0000FF  | empty    | #008000  | empty    |

The chart will be rendered as follows:

| Color attributes     | Dataset1      | Dataset2                    | Dataset3      | Dataset4      |
|----------------------|---------------|-----------------------------|---------------|---------------|
| **backgroundColors** | Theme color 1 | #FFFF00                     | Theme color 3 | Theme color 4 |
| **borderColors**     | #0000FF       | background color (#FFFF00)  | #008000       | Theme color 4 |